Working single-gene plot
The first plot works and can display 2 factors dynamically with any gene, faceted.
The space it takes up is small for the data (the plots are tiny), so I'll next attempt to enlarge their area dynamically in the UI somehow..

A critical feature I think we should add in the sidebarpanel is choosing the 2 factors to plot.