Automatic plot height adjustment
Got it to work by finding the number of rows in the ggplot, and multiplying by a number of pixels (350 currently).

I worked for hours trying to remove an overlap that occurred between the plots and the table. I didn't succeed, so I moved the table to it's own tab.